Route Protection: 컨텐츠 소유 확인후  편집/삭제

+++ controllers/videoController.js
@@ -42,8 +42,10 @@ export const postUpload = async (req, res) => {
         fileUrl: path,
         title,
         description,
+        creator: req.user.id,
     });
-
+    req.user.videos.push(newVideo.id);
+    req.user.save();
     res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -54,7 +56,7 @@ export const videoDetail = async (req, res) => {
     } = req;
 
     try {
-        const video = await Video.findById(id);
+        const video = await Video.findById(id).populate("creator");
         res.render("videoDetail", { pageTitle: video.title, video });
     } catch (error) {
         res.redirect(routes.home);
@@ -69,6 +71,9 @@ export const getEditVideo = async (req, res) => {
 
     try {
         const video = await Video.findById(id);
+        if (video.creator.toString() !== req.user.id) {
+            throw Error();
+        }
         res.render("editVideo", {
             pageTitle: `Edit Video ${video.title}`,
             video,
@@ -99,6 +104,10 @@ export const deleteVideo = async (req, res) => {
     } = req;
 
     try {
+        const video = await Video.findById(id);
+        if (video.creator.toString() !== req.user.id) {
+            throw Error();
+        }
         await Video.findOneAndRemove({ _id: id });
     } catch (error) {
         console.log(error);
